+++ superfields/src/main/java/org/vaadin/miki/superfields/object/ObjectField.java
@@ -7,6 +7,7 @@ import com.vaadin.flow.component.HasComponents;
 import com.vaadin.flow.component.HasStyle;
 import com.vaadin.flow.component.HasValue;
 import com.vaadin.flow.component.customfield.CustomField;
+import com.vaadin.flow.component.dependency.CssImport;
 import com.vaadin.flow.component.orderedlayout.FlexLayout;
 import com.vaadin.flow.function.SerializableSupplier;
 import org.slf4j.Logger;
@@ -16,6 +17,7 @@ import org.vaadin.miki.markers.WithHelperMixin;
 import org.vaadin.miki.markers.WithHelperPositionableMixin;
 import org.vaadin.miki.markers.WithIdMixin;
 import org.vaadin.miki.markers.WithLabelMixin;
+import org.vaadin.miki.markers.WithLabelPositionableMixin;
 import org.vaadin.miki.markers.WithValueMixin;
 import org.vaadin.miki.superfields.layouts.FlexLayoutHelpers;
 import org.vaadin.miki.superfields.object.builder.SimplePropertyComponentBuilder;
@@ -48,9 +50,10 @@ import java.util.stream.Collectors;
  * @author miki
  * @since 2022-05-16
  */
+@CssImport(value = "./styles/label-positions.css", themeFor = "vaadin-custom-field")
 public class ObjectField<T> extends CustomField<T>
         implements HasStyle, WithHelperMixin<ObjectField<T>>, WithHelperPositionableMixin<ObjectField<T>>,
-                   WithIdMixin<ObjectField<T>>, WithLabelMixin<ObjectField<T>>,
+                   WithIdMixin<ObjectField<T>>, WithLabelMixin<ObjectField<T>>, WithLabelPositionableMixin<ObjectField<T>>,
                    WithValueMixin<AbstractField.ComponentValueChangeEvent<CustomField<T>, T>, T, ObjectField<T>> {
 
     public static final SerializableSupplier<FlexLayout> DEFAULT_LAYOUT_PROVIDER = FlexLayoutHelpers::column;